+++ app/common/auth/password-reset.controller.js
