+++ app/common/common-module.js
@@ -1,18 +1,24 @@
 angular.module('ushahidi.common', [
-	'ushahidi.common.dropdown',
-	'ushahidi.common.accordion',
-	'ushahidi.common.offcanvas',
-	'ushahidi.common.modal',
-	'ushahidi.common.custom-on-change',
-	'ushahidi.common.file-upload',
-	'ushahidi.common.notification-slider',
-	'ushahidi.common.sticky-sidebar',
+    'ushahidi.common.dropdown',
+    'ushahidi.common.accordion',
+    'ushahidi.common.offcanvas',
+    'ushahidi.common.modal',
+    'ushahidi.common.custom-on-change',
+    'ushahidi.common.file-upload',
+    'ushahidi.common.notification-slider',
+    'ushahidi.common.sticky-sidebar',
     'ushahidi.common.chart'
 ])
 
-.service('Authentication', require('./services/authentication.js'))
-.service('Session', require('./services/session.js'))
+// Authentication
+.service('Authentication', require('./auth/authentication.service.js'))
+.service('Registration', require('./auth/registration.service.js'))
+.service('Session', require('./auth/session.service.js'))
+.service('PasswordReset', require('./auth/password-reset.service.js'))
+.config(require('./auth/authentication-interceptor.config.js'))
+.run(require('./auth/authentication-events.run.js'))
 
+// API Endpoint wrappers
 .service('ConfigEndpoint', require('./services/endpoints/config.js'))
 .service('UserEndpoint', require('./services/endpoints/user-endpoint.js'))
 .service('FormEndpoint', require('./services/endpoints/form.js'))
@@ -25,6 +31,7 @@ angular.module('ushahidi.common', [
 .service('MediaEndpoint', require('./services/endpoints/MediaEndpoint.js'))
 .service('MessageEndpoint', require('./services/endpoints/message.js'))
 
+// Other services
 .service('ViewHelper', require('./services/view-helper.js'))
 .service('Features', require('./services/features.js'))
 .service('Util', require('./services/util.js'))
@@ -35,17 +42,17 @@ angular.module('ushahidi.common', [
 .service('Maps', require('./services/maps.js'))
 .service('Geocoding', require('./services/geocoding.js'))
 .service('Languages', require('./services/languages.js'))
-.service('Registration', require('./services/registration.js'))
-.service('PasswordReset', require('./services/password-reset.js'))
 .service('IconManager', require('./services/icon-manager.js'))
 .service('FontAwesomeIcons', require('./services/endpoints/FontAwesomeIcons.js'))
 .service('ModalService', require('./services/modal.service.js'))
 
+// Controllers
 .controller('navigation', require('./controllers/navigation.js'))
 .controller('PageMetadata', require('./controllers/page-metadata.js'))
 .controller('notifier', require('./controllers/notifier.js'))
 .controller('intercom', require('./controllers/intercom.js'))
 
+// Global directives
 .directive('collectionSelector', require('./directives/collection-selector.js'))
 .directive('listingToolbar', require('./directives/list-toolbar.js'))
 .directive('iconPicker', require('./directives/iconpicker.js'))
@@ -64,8 +71,6 @@ angular.module('ushahidi.common', [
     }
 })
 
-
-.config(require('./configs/authentication-interceptor.js'))
 .config(require('./configs/locale-config.js'))
 .config(require('./configs/ui-bootstrap-template-decorators.js'))
 .config(require('./configs/cache-config.js'))
